Update logappend.py
Fixed append() to check for valid movement within the gallery. Currently, there are some small bugs with batch files. logappend will work correctly on batch files, but will output 'invalid' and return 255 if there is an error within the batch file, instead of '0' which it is supposed to do per the instructions.